libft update

+++ 42-Projets/libft/ft_strncmp.c
@@ -6,11 +6,11 @@
 /*   By: ewurstei <ewurstei@student.42quebec.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/02/25 14:29:16 by ewurstei          #+#    #+#             */
-/*   Updated: 2022/04/04 10:52:14 by ewurstei         ###   ########.fr       */
+/*   Updated: 2022/04/05 09:50:18 by ewurstei         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-int	ft_strncmp(const char *s1, const char *s2, unsigned int n)
+int	ft_strncmp(const char *s1, const char *s2, size_t n)
 {
 	unsigned int	i;
